Add files via upload
The file contains information on the depth-age relationship of the ExNGT 2012 shallow cores. For each of the cores the top and bottom depth of a year is given (and the related annual mean value in d18O).
Needs to be verified, if this is the same numbers (annual d18O per year) as was/is used in the analysis.